tests/Ranker.test.ts
// eloRanker.test.ts

import { Ranker } from "../src/ranker";
import { RankableItem, ComparisonResult } from "../src/types";

describe("EloRanker Class Tests", () => {
  let ranker: Ranker;

  beforeEach(() => {
    // Initialize the ranker with some items
    const items: RankableItem[] = [
      {
        id: "item1",
        initialRating: 1500,
        currentRating: 1500,
        comparisons: 0,
        wins: 0,
        losses: 0,
        ties: 0,
        stable: false,
      },
      {
        id: "item2",
        initialRating: 1500,
        currentRating: 1500,
        comparisons: 0,
        wins: 0,
        losses: 0,
        ties: 0,
        stable: false,
      },
      {
        id: "item3",
        initialRating: 1500,
        currentRating: 1500,
        comparisons: 0,
        wins: 0,
        losses: 0,
        ties: 0,
        stable: false,
      },
      {
        id: "item4",
        initialRating: 1500,
        currentRating: 1500,
        comparisons: 0,
        wins: 0,
        losses: 0,
        ties: 0,
        stable: false,
      },
    ];

    const config = {
      kFactor: 32,
      ratingChangeThreshold: 5,
      stableComparisonsThreshold: 10,
      minimumComparisons: 20,
      defaultInitialRating: 1500,
      minRating: 0,
    };

    ranker = new Ranker(items, config);
  });

  test("should initialize with correct item count", () => {
    expect(ranker.getItemCount()).toBe(4);
  });

  test("should add a new item correctly", () => {
    ranker.addItem("item5");
    expect(ranker.getItemCount()).toBe(5);
    expect(ranker.getItemStats("item5").currentRating).toBe(1500);
  });

  test("should not add an item with duplicate id", () => {
    expect(() => ranker.addItem("item1")).toThrowError(
      "Item with id item1 already exists"
    );
  });

  test("should remove an existing item", () => {
    ranker.removeItem("item1");
    expect(ranker.getItemCount()).toBe(3);
    expect(() => ranker.getItemStats("item1")).toThrowError(
      "Item with id item1 not found"
    );
  });

  test("should not remove a non-existent item", () => {
    expect(() => ranker.removeItem("nonexistent")).toThrowError(
      "Item with id nonexistent not found"
    );
  });

  test("should update ratings correctly after a win", () => {
    const result: ComparisonResult = {
      itemId1: "item1",
      itemId2: "item2",
      result: "win",
      timestamp: Date.now(),
    };

    ranker.addComparisonResult(result);

    const item1 = ranker.getItemStats("item1");
    const item2 = ranker.getItemStats("item2");

    expect(item1.currentRating).toBeGreaterThan(1500);
    expect(item2.currentRating).toBeLessThan(1500);
    expect(item1.wins).toBe(1);
    expect(item2.losses).toBe(1);
    expect(item1.comparisons).toBe(1);
    expect(item2.comparisons).toBe(1);
  });

  test("should update ratings correctly after a loss", () => {
    const result: ComparisonResult = {
      itemId1: "item1",
      itemId2: "item2",
      result: "loss",
      timestamp: Date.now(),
    };

    ranker.addComparisonResult(result);

    const item1 = ranker.getItemStats("item1");
    const item2 = ranker.getItemStats("item2");

    expect(item1.currentRating).toBeLessThan(1500);
    expect(item2.currentRating).toBeGreaterThan(1500);
    expect(item1.losses).toBe(1);
    expect(item2.wins).toBe(1);
    expect(item1.comparisons).toBe(1);
    expect(item2.comparisons).toBe(1);
  });

  test("should handle multiple comparisons and update rankings", () => {
    const results: ComparisonResult[] = [
      {
        itemId1: "item1",
        itemId2: "item2",
        result: "win",
        timestamp: Date.now(),
      },
      {
        itemId1: "item1",
        itemId2: "item3",
        result: "loss",
        timestamp: Date.now(),
      },
      {
        itemId1: "item2",
        itemId2: "item3",
        result: "win",
        timestamp: Date.now(),
      },
      {
        itemId1: "item2",
        itemId2: "item4",
        result: "tie",
        timestamp: Date.now(),
      },
      {
        itemId1: "item3",
        itemId2: "item4",
        result: "win",
        timestamp: Date.now(),
      },
      {
        itemId1: "item1",
        itemId2: "item4",
        result: "win",
        timestamp: Date.now(),
      },
    ];

    results.forEach((result) => ranker.addComparisonResult(result));

    const rankings = ranker.getRankings();

    console.log(
      "Final ratings:",
      rankings.map((item) => `${item.id}: ${item.currentRating}`)
    );

    // Check if the rankings are in the correct order
    expect(rankings[0].currentRating).toBeGreaterThanOrEqual(
      rankings[1].currentRating
    );
    expect(rankings[1].currentRating).toBeGreaterThanOrEqual(
      rankings[2].currentRating
    );
    expect(rankings[2].currentRating).toBeGreaterThanOrEqual(
      rankings[3].currentRating
    );

    // Check that item1 and item3 are in the top two positions
    expect(["item1", "item3"]).toContain(rankings[0].id);
    expect(["item1", "item3"]).toContain(rankings[1].id);
    expect(rankings[2].id).toBe("item2");
    expect(rankings[3].id).toBe("item4");

    // Check that the difference between the top two ratings is small
    expect(rankings[0].currentRating - rankings[1].currentRating).toBeLessThan(
      5
    );
  });

  test("should handle very large rating differences correctly", () => {
    ranker.addItem("highRatedItem", 2400);
    ranker.addItem("lowRatedItem", 1000);

    const result: ComparisonResult = {
      itemId1: "highRatedItem",
      itemId2: "lowRatedItem",
      result: "win",
      timestamp: Date.now(),
    };

    const highRatedItemBefore = ranker.getItemStats("highRatedItem");
    const lowRatedItemBefore = ranker.getItemStats("lowRatedItem");

    ranker.addComparisonResult(result);

    const highRatedItemAfter = ranker.getItemStats("highRatedItem");
    const lowRatedItemAfter = ranker.getItemStats("lowRatedItem");

    console.log("High rated item before:", highRatedItemBefore.currentRating);
    console.log("High rated item after:", highRatedItemAfter.currentRating);
    console.log("Low rated item before:", lowRatedItemBefore.currentRating);
    console.log("Low rated item after:", lowRatedItemAfter.currentRating);

    expect(highRatedItemAfter.currentRating).toBeGreaterThan(
      highRatedItemBefore.currentRating
    );
    expect(lowRatedItemAfter.currentRating).toBeLessThan(
      lowRatedItemBefore.currentRating
    );
    expect(
      highRatedItemAfter.currentRating - lowRatedItemAfter.currentRating
    ).toBeLessThan(1401);
  });

  test("should update ratings correctly after a tie", () => {
    const result: ComparisonResult = {
      itemId1: "item1",
      itemId2: "item2",
      result: "tie",
      timestamp: Date.now(),
    };

    ranker.addComparisonResult(result);

    const item1 = ranker.getItemStats("item1");
    const item2 = ranker.getItemStats("item2");

    expect(item1.currentRating).toBe(1500);
    expect(item2.currentRating).toBe(1500);
    expect(item1.ties).toBe(1);
    expect(item2.ties).toBe(1);
    expect(item1.comparisons).toBe(1);
    expect(item2.comparisons).toBe(1);
  });

  test("should not allow comparing an item with itself", () => {
    const result: ComparisonResult = {
      itemId1: "item1",
      itemId2: "item1",
      result: "win",
      timestamp: Date.now(),
    };

    expect(() => ranker.addComparisonResult(result)).toThrowError(
      "Cannot compare an item with itself"
    );
  });

  test("should not allow ratings to drop below minRating", () => {
    // Create a scenario where an item's rating would drop below minRating if not constrained
    const losingStreakResult: ComparisonResult = {
      itemId1: "item1",
      itemId2: "item2",
      result: "loss",
      timestamp: Date.now(),
    };

    // Simulate a long losing streak
    for (let i = 0; i < 100; i++) {
      ranker.addComparisonResult(losingStreakResult);
    }

    const item1 = ranker.getItemStats("item1");
    expect(item1.currentRating).toBeGreaterThanOrEqual(0);
  });

  test("should correctly identify stable items", () => {
    // Simulate comparisons to reach stability
    const result: ComparisonResult = {
      itemId1: "item1",
      itemId2: "item2",
      result: "win",
      timestamp: Date.now(),
    };

    for (let i = 0; i < 25; i++) {
      ranker.addComparisonResult(result);
    }

    const item1 = ranker.getItemStats("item1");
    expect(item1.stable).toBe(true);

    const item2 = ranker.getItemStats("item2");
    expect(item2.stable).toBe(true);
  });

  test("should return correct progress", () => {
    expect(ranker.getProgress()).toBe(0);

    // Simulate comparisons to stabilize some items
    const result: ComparisonResult = {
      itemId1: "item1",
      itemId2: "item2",
      result: "win",
      timestamp: Date.now(),
    };

    for (let i = 0; i < 25; i++) {
      ranker.addComparisonResult(result);
    }

    expect(ranker.getProgress()).toBe(0.5);
  });

  test("should return null when no comparisons are needed", () => {
    // Stabilize all items
    const results: ComparisonResult[] = [];

    // Generate enough comparisons to stabilize all items
    for (let i = 0; i < 25; i++) {
      results.push({
        itemId1: "item1",
        itemId2: "item2",
        result: "win",
        timestamp: Date.now(),
      });
      results.push({
        itemId1: "item3",
        itemId2: "item4",
        result: "win",
        timestamp: Date.now(),
      });
    }

    results.forEach((result) => ranker.addComparisonResult(result));

    expect(ranker.getNextComparison()).toBeNull();
  });

  test("should provide correct item statistics", () => {
    const result: ComparisonResult = {
      itemId1: "item1",
      itemId2: "item2",
      result: "win",
      timestamp: Date.now(),
    };

    ranker.addComparisonResult(result);

    const stats = ranker.getItemStats("item1");
    expect(stats.wins).toBe(1);
    expect(stats.losses).toBe(0);
    expect(stats.ties).toBe(0);
    expect(stats.comparisons).toBe(1);
  });

  test("should return a pair of items for the next comparison", () => {
    const nextComparison = ranker.getNextComparison();
    expect(nextComparison).not.toBeNull();
    expect(nextComparison).toHaveLength(2);
    expect(nextComparison![0]).not.toBe(nextComparison![1]);
  });

  test("should throw an error when adding a comparison result with invalid item IDs", () => {
    const result: ComparisonResult = {
      itemId1: "item1",
      itemId2: "nonexistent",
      result: "win",
      timestamp: Date.now(),
    };

    expect(() => ranker.addComparisonResult(result)).toThrowError(
      "One or both items not found"
    );
  });

  test("should correctly reset stability when new comparisons affect stable items", () => {
    // Stabilize item1 and item2
    const result: ComparisonResult = {
      itemId1: "item1",
      itemId2: "item2",
      result: "win",
      timestamp: Date.now(),
    };

    for (let i = 0; i < 25; i++) {
      ranker.addComparisonResult(result);
    }

    const item1 = ranker.getItemStats("item1");
    expect(item1.stable).toBe(true);

    // Introduce a new result that causes a significant rating change
    const significantResult: ComparisonResult = {
      itemId1: "item1",
      itemId2: "item3",
      result: "loss",
      timestamp: Date.now(),
    };

    ranker.addComparisonResult(significantResult);

    // Since the implementation marks items as stable permanently, item1 remains stable
    // If the design intended for stability to be re-evaluated, additional logic would be required
    expect(ranker.getItemStats("item1").stable).toBe(true);
  });

  test("should maintain sorted rankings after each comparison", () => {
    // Perform several comparisons
    const results: ComparisonResult[] = [
      {
        itemId1: "item1",
        itemId2: "item2",
        result: "win",
        timestamp: Date.now(),
      },
      {
        itemId1: "item1",
        itemId2: "item3",
        result: "win",
        timestamp: Date.now(),
      },
      {
        itemId1: "item1",
        itemId2: "item4",
        result: "win",
        timestamp: Date.now(),
      },
      {
        itemId1: "item2",
        itemId2: "item3",
        result: "loss",
        timestamp: Date.now(),
      },
      {
        itemId1: "item2",
        itemId2: "item4",
        result: "loss",
        timestamp: Date.now(),
      },
    ];

    results.forEach((result) => ranker.addComparisonResult(result));

    const rankings = ranker.getRankings();

    // Check that rankings are sorted in descending order of currentRating
    for (let i = 0; i < rankings.length - 1; i++) {
      expect(rankings[i].currentRating).toBeGreaterThanOrEqual(
        rankings[i + 1].currentRating
      );
    }
  });

  test("should correctly handle tie scenarios in ratings", () => {
    const result: ComparisonResult = {
      itemId1: "item1",
      itemId2: "item2",
      result: "tie",
      timestamp: Date.now(),
    };

    ranker.addComparisonResult(result);

    const item1 = ranker.getItemStats("item1");
    const item2 = ranker.getItemStats("item2");

    expect(item1.currentRating).toBe(1500);
    expect(item2.currentRating).toBe(1500);
  });

  test("should correctly calculate progress when there are no items", () => {
    const emptyRanker = new Ranker([], {});
    expect(emptyRanker.getProgress()).toBe(1);
  });

  test("should maintain correct order of items with equal ratings", () => {
    ranker.addItem("equalItem1", 1500);
    ranker.addItem("equalItem2", 1500);

    const rankings = ranker.getRankings();
    const equalItem1Index = rankings.findIndex(
      (item) => item.id === "equalItem1"
    );
    const equalItem2Index = rankings.findIndex(
      (item) => item.id === "equalItem2"
    );

    expect(Math.abs(equalItem1Index - equalItem2Index)).toBe(1);
  });

  test("should correctly handle a long series of alternating wins and losses", () => {
    const initialRating1 = ranker.getItemStats("item1").currentRating;
    const initialRating2 = ranker.getItemStats("item2").currentRating;

    for (let i = 0; i < 100; i++) {
      const result: ComparisonResult = {
        itemId1: "item1",
        itemId2: "item2",
        result: i % 2 === 0 ? "win" : "loss",
        timestamp: Date.now(),
      };
      ranker.addComparisonResult(result);
    }

    const finalRating1 = ranker.getItemStats("item1").currentRating;
    const finalRating2 = ranker.getItemStats("item2").currentRating;

    expect(Math.abs(finalRating1 - initialRating1)).toBeLessThan(100);
    expect(Math.abs(finalRating2 - initialRating2)).toBeLessThan(100);
  });

  test("should correctly apply custom initial ratings", () => {
    ranker.addItem("customItem", 2000);
    expect(ranker.getItemStats("customItem").currentRating).toBe(2000);
    expect(ranker.getItemStats("customItem").initialRating).toBe(2000);
  });

  test("should throw an error when trying to get stats for a non-existent item", () => {
    expect(() => ranker.getItemStats("nonExistentItem")).toThrowError(
      "Item with id nonExistentItem not found"
    );
  });

  test("should correctly handle rating changes near the minimum rating", () => {
    ranker.addItem("lowRatedItem", 10);

    const result: ComparisonResult = {
      itemId1: "lowRatedItem",
      itemId2: "item1",
      result: "loss",
      timestamp: Date.now(),
    };

    ranker.addComparisonResult(result);

    const lowRatedItem = ranker.getItemStats("lowRatedItem");
    expect(Math.round(lowRatedItem.currentRating)).toBe(10);
  });

  test("should return correct getAllItems result", () => {
    const allItems = ranker.getAllItems();
    expect(allItems.length).toBe(ranker.getItemCount());
    expect(allItems.every((item) => item.id && item.currentRating)).toBe(true);
  });

  test("should handle a large number of items correctly", () => {
    for (let i = 0; i < 1000; i++) {
      ranker.addItem(`largeItem${i}`);
    }

    expect(ranker.getItemCount()).toBe(1004); // 1000 new items + 4 initial items
    expect(() => ranker.getNextComparison()).not.toThrow();
    expect(ranker.getRankings().length).toBe(1004);
  });
});
